requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# PDF Processing
pdfplumber>=0.10.3
pymupdf>=1.23.0
pdf2image>=1.16.3
camelot-py[cv]>=0.11.0
tabula-py>=2.7.0

# OCR
pytesseract>=0.3.10
PaddleOCR>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0

# Vector Database
faiss-cpu>=1.7.0
chromadb>=0.4.18

# LLM Local
ollama>=0.1.0
litellm>=1.0.0

# UI
streamlit>=1.28.0
gradio>=3.45.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
click>=8.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0